BotGruppe/pyproject.toml
[tool.poetry]
name = "GruppeBot"
version = "0.1.0"
description = ""
authors = ["arfuk9038@example.com"]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.0"
apscheduler = "^3.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"